requirements.txt
# Telegram bot
python-telegram-bot[job-queue]>=20.7
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0

# Web interface
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.2.0

# Price data
yfinance>=0.2.40
openpyxl>=3.1.0
